Combine Prompt.cpp  file with Precheck (#42)
* Combine Prompt file with Precheck file

* Update DatabasePrecheck.cpp

* Update DatabasePrecheck.h

* Update DatabasePrecheck.cpp
