+++ WeAlumni/WeAlumni.vcxproj.filters
@@ -114,15 +114,10 @@
     <ClCompile Include="StartPage2.cpp">
       <Filter>Source Files\System</Filter>
     </ClCompile>
-    <ClCompile Include="DatabasePrecheck.cpp">
-      <Filter>Source Files\Database</Filter>
-    </ClCompile>
-    <ClCompile Include="DatabasePrecheckPrompt.cpp">
-      <Filter>Source Files\Database</Filter>
-    </ClCompile>
     <ClCompile Include="LoginPage.cpp">
       <Filter>Source Files\System</Filter>
     </ClCompile>
+    <ClCompile Include="DatabasePrecheck.cpp">
     <ClCompile Include="DBImportExport.cpp">
       <Filter>Source Files\Database</Filter>
     </ClCompile>
@@ -188,15 +183,10 @@
     <ClInclude Include="SysForgotPassword.h">
       <Filter>Header Files\System</Filter>
     </ClInclude>
-    <ClInclude Include="DatabasePrecheck.h">
-      <Filter>Header Files\Database</Filter>
-    </ClInclude>
-    <ClInclude Include="DatabasePrecheckPrompt.h">
-      <Filter>Header Files\Database</Filter>
-    </ClInclude>
     <ClInclude Include="LoginPage.h">
       <Filter>Header Files\System</Filter>
     </ClInclude>
+    <ClInclude Include="DatabasePrecheck.h">
     <ClInclude Include="DBImportExport.h">
       <Filter>Header Files\Database</Filter>
     </ClInclude>
